feat!: move to straight tiff naming
