+++ packages/core/src/index.ts
@@ -1,5 +1,5 @@
-export { CogTiffImage } from './cog.tiff.image.js';
-export { CogTiff } from './cog.tiff.js';
+export { CogTiffImage } from './tiff.image.js';
+export { CogTiff } from './tiff.js';
 export { TiffEndian } from './const/tiff.endian.js';
 export { TiffCompression, TiffMimeType } from './const/tiff.mime.js';
 export { TiffTag, TiffTagGeo, TiffTagGeoValueLookup, TiffTagValueLookup } from './const/tiff.tag.id.js';
